requirements-kuzu-0.0.11.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements-kuzu-0.0.11.in
#
bcrypt==4.1.2
    # via paramiko
cffi==1.16.0
    # via
    #   cryptography
    #   pynacl
cryptography==42.0.2
    # via paramiko
faker==23.2.0
    # via -r requirements-kuzu-0.0.11.in
kuzu==0.0.11
    # via -r requirements-kuzu-0.0.11.in
numpy==1.26.4
    # via
    #   -r requirements-kuzu-0.0.11.in
    #   pandas
    #   pyarrow
pandas==2.2.0
    # via -r requirements-kuzu-0.0.11.in
paramiko==3.4.0
    # via -r requirements-kuzu-0.0.11.in
prettytable==3.9.0
    # via -r requirements-kuzu-0.0.11.in
pyarrow==15.0.0
    # via -r requirements-kuzu-0.0.11.in
pycparser==2.21
    # via cffi
pynacl==1.5.0
    # via paramiko
python-dateutil==2.8.2
    # via
    #   faker
    #   pandas
python-dotenv==1.0.1
    # via -r requirements-kuzu-0.0.11.in
pytz==2024.1
    # via pandas
six==1.16.0
    # via python-dateutil
tqdm==4.66.2
    # via -r requirements-kuzu-0.0.11.in
tzdata==2024.1
    # via pandas
wcwidth==0.2.13
    # via prettytable
